requirements_dev.txt
-r requrements.txt

django-debug-toolbar==1.5
factory-boy==2.12.0
ipython==6.5.0
pytest==5.1.0
pytest-django==3.5.1
